move debug_timing to utils
